move ContentAnalyticsActionEvent to the base-tool module
