+++ ui/base-tool/src/main/java/org/cru/godtools/base/tool/analytics/model/ContentAnalyticsActionEvent.kt
@@ -1,4 +1,4 @@
-package org.cru.godtools.tract.analytics.model
+package org.cru.godtools.base.tool.analytics.model
 
 import android.os.Bundle
 import androidx.annotation.VisibleForTesting
@@ -6,7 +6,7 @@ import org.cru.godtools.analytics.model.AnalyticsActionEvent
 import org.cru.godtools.analytics.model.AnalyticsSystem
 import org.cru.godtools.xml.model.AnalyticsEvent
 
-class ContentAnalyticsActionEvent(@VisibleForTesting internal val event: AnalyticsEvent) :
+class ContentAnalyticsActionEvent(@VisibleForTesting val event: AnalyticsEvent) :
     AnalyticsActionEvent(action = event.action.orEmpty()) {
     override fun isForSystem(system: AnalyticsSystem) = event.isForSystem(system)
