+++ ui/tract-renderer/src/main/java/org/cru/godtools/tract/ui/controller/BaseController.kt
@@ -17,8 +17,8 @@ import kotlinx.coroutines.launch
 import org.ccci.gto.android.common.androidx.lifecycle.ImmutableLiveData
 import org.ccci.gto.android.common.db.findLiveData
 import org.cru.godtools.base.model.Event
+import org.cru.godtools.base.tool.analytics.model.ContentAnalyticsActionEvent
 import org.cru.godtools.model.TrainingTip
-import org.cru.godtools.tract.analytics.model.ContentAnalyticsActionEvent
 import org.cru.godtools.xml.model.AnalyticsEvent
 import org.cru.godtools.xml.model.Base
 import org.cru.godtools.xml.model.layoutDirection
